+++ gui/extension/CHANGELOG.md
@@ -11,6 +11,7 @@
 - BUG#36173373 Parent class field not accessible in the child class in the TypeScript SDK
 - BUG#36211402 Unexpected compilation error when building the VSCode Extension
 - BUG#35150090 Missing required options on "mrs.add.schema()" documentation
+- BUG#36204648 Syntax error when query filter mixes explicit OR and implicit AND operators
 
 ## Changes in 1.14.2+8.1.1